♻️ ChatPannel에서 InputBox 분리
